requirements.txt
fastapi
uvicorn
slowapi
duckduckgo-search
groq
python-dotenv
